pyproject.toml
[tool.poetry]
name = "crew-haiku"
version = "0.1.0"
description = "A CrewAI project for generating haikus using local LLM"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "*.py" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = "^0.95.0"
langchain-community = "^0.3.14"
langchain-ollama = "^0.2.2"
langchain-openai = "^0.3.0"
python-dotenv = "^1.0.1"
flask = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^6.0"
